renamed events_new.py to events.py; delete old events.py
